requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot==21.10

# Database
sqlalchemy==2.0.35

# Utils
python-dotenv==1.0.0
pydantic==2.8.0
pydantic-settings==2.4.0

# CORS
fastapi-cors==0.0.6

# Optional
httpx==0.27.0